requirements.txt
# Core AI / CV
ultralytics>=8.2.0          # YOLOv8 — includes PyTorch dependency
opencv-python-headless>=4.9.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.3.0

# Database
sqlalchemy>=2.0.0

# Config
pyyaml>=6.0.1

# Optional: GPU acceleration (uncomment if CUDA is available)
# torch>=2.2.0+cu121
# torchvision>=0.17.0+cu121
